Comparar carrito vs stock en front-end
