+++ Marketplace/app.js
@@ -47,10 +47,12 @@ const rutaProducto = require("./routers/product");
 const rutaUsuario = require("./routers/user");
 const rutaAPI = require("./routers/api");
 const rutaCarrito = require("./routers/carrito");
+const rutaVenta = require("./routers/venta");
 const rutaMain = require("./routers/main");
 // **************** Rutas *******************
 app.use("/producto", rutaProducto);
 app.use("/usuario", rutaUsuario);
 app.use("/api", rutaAPI);
 app.use("/carrito", rutaCarrito);
+app.use("/venta", rutaVenta);
 app.use("/", rutaMain);
