+++ Marketplace/public/javascript/carrito.js
@@ -1,5 +1,5 @@
-window.addEventListener("load", () => {
-	// Declarar lasvariables
+window.addEventListener("load", async () => {
+	// Declarar las variables
 	let cantidad = document.querySelectorAll("#cantidad");
 	let precio = document.querySelectorAll("#precio");
 	let importe = document.querySelector("#importe");
@@ -7,12 +7,28 @@ window.addEventListener("load", () => {
 	let carritoID = document.querySelectorAll("#carritoID");
 	let comprar = document.querySelector("#comprar");
 
+	// Obtener el stock de cada producto
+	let api = await fetch("/api/productos")
+	.then((n) => n.json())
+	.then(n => n.products)
+	let productos = document.querySelectorAll("#productoID");
+	let stock = []
+	for (n of productos) {
+		ID = n.innerHTML
+		stockDisponible = api.find((m) => m.id == ID).stock;
+		stock.push(stockDisponible);
+	}
 	// Rutinas por cada carrito
 	for (let i = 0; i < cantidad.length; i++) {
 		// Cambios en la cantidad
 		cantidad[i].addEventListener("input", () => {
 			cant = parseInt(cantidad[i].value);
-			cant < 0 ? (cant = 0) : "";
+			if (cant < 0) {
+				cant = 0
+			} else
+			if (cant > stock[i]) {
+				cant = stock[i]
+			}
 			actualizar(cant, i);
 			comprar.classList.add("ocultar")
 		});
