+++ Marketplace/repositories/ventasRepository.js
@@ -1,23 +1,8 @@
 const db = require("../database/models");
 const encabezado = db.encabezadoVenta;
 const detalle = db.detalleVenta;
-const compra = db.Carrito;
 
 module.exports = {
-	ImporteCompra: async (usuarioID) => {
-		let carrito = await compra.findAll({
-			include: ["producto"],
-			where: { usuario_id: usuarioID },
-		});
-		let acumulador = 0;
-		for (n of carrito) {
-			cant = n.cantidad;
-			precio = n.producto.precio;
-			acumulador = acumulador + cant * precio;
-		}
-		return acumulador;
-	},
-
 	AgregarCabecera: async (usuarioID, importe) => {
 		if (await encabezado.findAll().then(n => (n.length == 0))) {
 			var numeroFC = 1
