+++ Marketplace/views/varios/carrito.ejs
@@ -22,7 +22,7 @@
 						</div>
 						<!--CONTINUAR Y PAGAR-->
 						<div id="continuar-y-pagar">
-							<button id="comprar"><a href="/carrito/comprar">Comprar</a></button>
+							<button id="comprar"><a href="/venta/kickOff">Comprar</a></button>
 							<button type="submit">Guardar cambios</button>
 						</div>
 					<% } %>
@@ -36,7 +36,7 @@
 					<% for (let i = 0; i < cantRegistros; i++) { %>
 						<%- include('carrito-registro', {
 							numRegistro: i,
-							carritoID: carritos[i].id,
+							carrito: carritos[i],
 							cantidad: carritos[i].cantidad,
 							producto: carritos[i].producto, 
 							imagen: carritos[i].imagen,
